Code cleanup in "bins".
